requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
babel==2.17.0
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
clickhouse-connect==0.8.18
courlan==1.3.2
cryptography==45.0.6
dateparser==1.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.7.0
h11==0.16.0
hf-xet==1.1.9
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jusText==3.0.2
langchain==0.3.27
langchain-core==0.3.75
langchain-text-splitters==0.3.9
langsmith==0.4.20
loguru==0.7.3
lxml==5.4.0
lxml_html_clean==0.4.2
lz4==4.4.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.6
orjson==3.11.3
packaging==25.0
pandas==2.3.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
prometheus_client==0.23.1
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.4
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2025.7.34
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentence-transformers==5.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tld==0.13.1
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.56.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
zstandard==0.24.0
